requirements.txt
Django==5.2.4
djangorestframework==3.16.1
psycopg2==2.9.10
python-dotenv==1.1.1
django-debug-toolbar==6.0.0
pillow==11.3.0
requests==2.32.4
pyTelegramBotAPI==4.28.0
cachetools==6.1.0